wasm poc cas client compiles (#282)
This PR makes the cas_client crate compile into WASM (at least with
wasm-pack). It separates out download & upload traits and functions when
necessary to then target_family block the downloads path (easy way out).
There's other generally necessary changes particularly having to do with
async_trait that are added.

---------
