+++ cas_client/Cargo.toml
@@ -9,7 +9,7 @@ strict = []
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
-cas_types = { version = "0.1.0", path = "../cas_types" }
+cas_types = { path = "../cas_types" }
 cas_object = { path = "../cas_object" }
 chunk_cache = { path = "../chunk_cache" }
 error_printer = { path = "../error_printer" }
@@ -19,7 +19,7 @@ merkledb = { path = "../merkledb" }
 mdb_shard = { path = "../mdb_shard" }
 merklehash = { path = "../merklehash" }
 xet_threadpool = { path = "../xet_threadpool" }
-deduplication = {path = "../deduplication" }
+deduplication = { path = "../deduplication" }
 thiserror = "2.0"
 tokio = { version = "1.44", features = ["sync"] }
 async-trait = "0.1.9"
@@ -27,17 +27,16 @@ anyhow = "1"
 http = "1.1.0"
 tempfile = "3.13.0"
 tracing = "0.1.31"
-bytes = "1"
-itertools = "0.10"
-reqwest = { version = "0.12.7", features = ["json", "stream"] }
-reqwest-middleware = "0.3.3"
+reqwest = { version = "0.12.15", features = ["json", "stream"] }
+reqwest-middleware = "0.4.1"
 url = "2.5.2"
-reqwest-retry = "0.6.1"
-heed = "0.11"
+reqwest-retry = "0.7.0"
 futures = "0.3.31"
 derivative = "2.2.0"
 
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+heed = "0.11"
+
 [dev-dependencies]
-rand = "0.8.5"
 httpmock = "0.7.0"
 tracing-test = { version = "0.2.5", features = ["no-env-filter"] }
