+++ cas_client/src/lib.rs
@@ -3,16 +3,25 @@
 pub use chunk_cache::{CacheConfig, CHUNK_CACHE_SIZE_BYTES};
 pub use http_client::{build_auth_http_client, build_http_client, RetryConfig};
 use interface::RegistrationClient;
-pub use interface::{Client, FileProvider, OutputProvider, Reconstructable, ReconstructionClient, UploadClient};
+pub use interface::{Client, UploadClient};
+#[cfg(not(target_family = "wasm"))]
+pub use interface::{Reconstruct, ReconstructionClient};
+#[cfg(not(target_family = "wasm"))]
 pub use local_client::LocalClient;
+#[cfg(not(target_family = "wasm"))]
+pub use output_provider::{FileProvider, OutputProvider};
 pub use remote_client::RemoteClient;
 
 pub use crate::error::CasClientError;
 pub use crate::interface::ShardClientInterface;
 
+#[cfg(not(target_family = "wasm"))]
 mod download_utils;
 mod error;
 mod http_client;
 mod interface;
+#[cfg(not(target_family = "wasm"))]
 mod local_client;
+#[cfg(not(target_family = "wasm"))]
+mod output_provider;
 pub mod remote_client;
