+++ cas_client/src/local_client.rs
@@ -21,7 +21,8 @@ use tracing::{debug, error, info, warn};
 use utils::progress::ProgressUpdater;
 
 use crate::error::{CasClientError, Result};
-use crate::interface::{OutputProvider, ShardDedupProber, UploadClient};
+use crate::interface::{ShardDedupProbe, UploadClient};
+use crate::output_provider::OutputProvider;
 use crate::{Client, ReconstructionClient, RegistrationClient, ShardClientInterface};
 
 pub struct LocalClient {
@@ -354,7 +355,7 @@ impl FileReconstructor<CasClientError> for LocalClient {
 }
 
 #[async_trait]
-impl ShardDedupProber for LocalClient {
+impl ShardDedupProbe for LocalClient {
     async fn query_for_global_dedup_shard(
         &self,
         _prefix: &str,
