+++ data/Cargo.toml
@@ -50,7 +50,7 @@ serde_json = "1.0.133"
 
 # Need to specify this as optional to allow the openssl/vendored option below
 openssl = { version = "0.10", features = [], optional = true }
-reqwest-middleware = "0.3.3"
+reqwest-middleware = "0.4.1"
 chrono = "0.4.39"
 
 [target.'cfg(not(windows))'.dependencies]
@@ -59,10 +59,10 @@ openssl = "0.10"
 # use embedded webpki root certs for MacOS as native certs take a very long time
 # to load, which affects startup time significantly
 [target.'cfg(macos)'.dependencies]
-reqwest = { version = "0.11.4", features = ["json", "webpki-roots"] }
+reqwest = { version = "0.12.15", features = ["json", "rustls-tls-webpki-roots"] }
 
 [target.'cfg(not(macos))'.dependencies]
-reqwest = { version = "0.11.4", features = ["json"] }
+reqwest = { version = "0.12.15", features = ["json"] }
 
 # Windows doesn't support assembly for compilation
 [target.'cfg(not(target_os = "windows"))'.dependencies]
