+++ data/src/remote_client_interface.rs
@@ -1,7 +1,7 @@
 use std::sync::Arc;
 
 pub use cas_client::Client;
-use cas_client::{LocalClient, RemoteClient};
+use cas_client::RemoteClient;
 use xet_threadpool::ThreadPool;
 
 use crate::configurations::*;
@@ -24,6 +24,13 @@ pub(crate) fn create_remote_client(
             config.shard_config.cache_directory.clone(),
             dry_run,
         ))),
-        Endpoint::FileSystem(ref path) => Ok(Arc::new(LocalClient::new(path, None)?)),
+        Endpoint::FileSystem(ref path) => {
+            #[cfg(not(target_family = "wasm"))]
+            {
+                Ok(Arc::new(cas_client::LocalClient::new(path, None)?))
+            }
+            #[cfg(target_family = "wasm")]
+            unimplemented!("Local file system access is not supported in WASM builds")
+        },
     }
 }
