+++ hf_xet/Cargo.toml
@@ -29,7 +29,7 @@ tracing-subscriber = { version = "0.3", features = [
     "env-filter",
 ] }
 bipbuffer = "0.1"
-reqwest = "0.11.27"
+reqwest = "0.12.15"
 serde = { version = "1.0.215", features = ["derive"] }
 serde_json = "1.0.133"
 lazy_static = "1.5"
