+++ mdb_shard/src/shard_file_reconstructor.rs
@@ -1,9 +1,9 @@
-use async_trait::async_trait;
 use merklehash::MerkleHash;
 
 use crate::file_structs::MDBFileInfo;
 
-#[async_trait]
+#[cfg_attr(not(target_arch = "wasm32"), async_trait::async_trait)]
+#[cfg_attr(target_arch = "wasm32", async_trait::async_trait(?Send))]
 pub trait FileReconstructor<E> {
     /// Returns a pair of (file reconstruction information,  maybe shard ID)
     /// Err(_) if an error occured
